Throwing error based on best practices - removing create bc @pgmaganty handling it - renaming file name
